# plan: remove parents totally