--- a/plan/join_reorder.go
+++ b/plan/join_reorder.go
@@ -190,7 +190,7 @@ func (e *joinReOrderSolver) newJoin(lChild, rChild LogicalPlan) *LogicalJoin {
 		reordered: true,
 	}.init(e.ctx)
 	join.SetSchema(expression.MergeSchema(lChild.Schema(), rChild.Schema()))
-	setParentAndChildren(join, lChild, rChild)
+	join.SetChildren(lChild, rChild)
 	return join
 }