remove parents list.